First steps towards adding multi-entry functions to the SSA IL
This is just a structural change to the datatypes in ssa-tree, with a
few behavioral changes implemented to make the new SSA structure
type-check on itself (but all of the transformations are broken by this
change).

This change introduces entry points, which are posessed by functions.
Next on the to-do list will be removing arguments from functions, since
that now belongs to entry points.