Preparation for building SXML -> ME_SSA conversion
As with the changes to SSA_TREE, we will keep the existing files where
they are and copy the exising pipeline files.  This way, we can still do
a build with the old pipeline, but when we switch to the new pipeline
we will still have all of the history from the old pipeline.

When there is a conflict for a structure name, new structures for
mutli-entry functions are prefixed with ME_.