Copy SSA passes that are necessary for the Ssa functor
To connect the closure converter to the new, multi-entry SSA, the Ssa
functor needs to be modified so that it exports the ME_SSA_TREE.  To
modify the Ssa functor, the following passes will also need to be
modified to conform to the new, multi-entry SSA:

- Simplify
- Restore
- Shrink
- Prepasses
- TypeCheck
- Analyze
- DirectExp

The Simplify pass (minimally) depends on:

- PolyEqual
- PolyHash
- SSA_TRANSFORM

The Simplify pass has other dependencies, but they are optional and will
be converted later.

The actual changes will appear in a separate commit, because I would
like to separate simple copying (this commit) from the actual changes in
order to make reading through the diffs more manageable.  When the
transition to multi-entry functions is complete, these files will be
deleted anyways.