Convert the SSA restore functor to the multi-entry SSA IL
The major change in this commit is how extra entry blocks are added
before calculating the phi functions.  Instead of adding a new entry
block at the beginning of a function, we add a new entry block for each
function entry.  After the phi functions have been added, the entry
blocks are rebuilt again.