Convert SSA Simplify to the multi-entry SSA IL (partially)
This change updates SSA Simplify to work with multi-entry functions.
Since Simplify primarily delegates to other passes, this was a matter of
commenting out all but the most essential passes.  The two mandatory
passes that weren't commented out are PolyEqual and PolyHash.

PolyEqual and PolyHash were converted in similar ways.  The addition of
entry points to functions means that callers must know both the function
being called and the function entry being called.  So, all that needed
to be done was make sure that both were being passed around when they
were being added to the property lists.