Export a multi-entry SSA functor
A new functor, MeSsa now exports a multi-entry SSA IL.  This is the
culmination of:

ddd66b1 Convert SSA Simplify to the multi-entry SSA IL (partially)
48134ba Convert the SSA restore functor to the multi-entry SSA IL
4910911 Convert the Shrink functor to the multi-entry SSA IL
24722bb Convert the SSA PrePasses functor to the multi-entry IL
251bd86 Convert SSA type checker to multi-entry SSA (partially)
70cf8a7 Convert SSA Analyze to the multi-entry SSA IL
7ba79bb Convert direct expressions to the multi-entry SSA IL
ca73cbd Copy SSA passes that are necessary for the Ssa functor

As noted in those commits, this is a step towards getting a minimal
pipeline supporting multi-entry functions.  The next step is to connect
the closure converter to this new functor.