SSA layout: add the function name when printing calls
The original intention for the 'entry' field of the Transfer.Call
constructor was that the 'entry' would be the main way that a function was
referred to.  Leaving 'func' in was just a way to get the name of the
parent function for some passes.

Since then, I've noticed myself creating function entries with names
that aren't very useful on their own, such as "default".  They're still
fine as labels, since they'll be alpha-renamed to be unique, but when
printing out an SSA program it's a lot more useful to see the name of
the function that is being called, instead of just the entry point.