Fix dominatorForest calculation for SSA2
Copy 60efb65 "Fix dominatorForest calculation." to the SSA2
implementation of the dominatorForest.

It's also become apparent that the fakeRoot and fakeBlock just float
around after the dominator forest has been built, since they are a part
of the graph.  Another pass will have to be made to actually remove
these from the graph, in a later commit.