Convert SSA2 Simplify to the multi-entry SSA2 IL (partially)
As in ddd66b11 (which was for SSA Simplify), the "conversion" is simply
disabling optional SSA2 passes.  In this case, there were no mandatory
passes so all of them have been disabled.

The disabled passes will be added back in once the pipeline for
multi-entry functions is complete and the regression test suite can run.