Port ssa2-to-rssa to the multi-entry function ILs
This conversion was mostly straightforward, since only functions and
transfers have changed.

One part of this change makes me uneasy: the conversion for 'main'
seemed a bit out of the ordinary, since it creates a new SSA2 function,
which points to the existing SSA2 'main', then transforms the new
function into RSSA, saving the new function as RSSA's 'main'.  Keeping
track of which labels needed to go into the new entry point and the new
transfer was confusing.  I think this is correct, but unfortunately, I
won't be able to test it until the rest of the backend can be added to
the main compilation pipeline.