Add two transformations to the list of minimal SSA transformations
Running the regression test suite with the '-opt-passes minimal' flag
generated executables that made it through the compiler, but didn't
accomplish much more than going in an infinite loop, allocating memory.
I haven't determined the root cause of that behavior, but it does not
happen when the inliner and constant propagation passes are run.  The
newly mandatory passes are:

inlineLeaf1
constantPropagation

Other failures still happen when running with '-opt-passes minimal':
some of the regression tests end up with case expressions that are
exhaustive but still have a default case added.  The SSA type checker
aborts the compilation there, since the default case is unreachable.
Once that problem is discovered, a later fix should make sure that
'-opt-passes minimal' *really* produces working executable.  In the mean
time, having most of the regression tests pass is enough of a baseline
to use for testing the multi-entry function pipeline.