Copy files for the single-entry backend
Copy the files that will need to be modified to get the backend
converted to using multi-entry functions.  Notably, the Machine IL
doesn't need to be changed, so it doesn't appear here.  Everything that
converted RSSA -> Machine IL does need to be updated, since the
multi-entry RSSA will have to be translated to the Machine IL
differently.

These files will be removed when the multi-entry pipeline is fully set
up (passing all regression tests).