Convert exception handlers for multi-entry functions
Stack slots for exceptions are allocated at each entry point.  Since the
exceptions that a function handles are global for a function, it's just
a matter of making sure that each function entry has the same exception
stack slots set up.

Exception history seems to be managed with the rest of the profiling
information, which is not yet functional.  So, while exception handlers
work with this change, generating traces won't.