Use correct entry point for RSSA main function
When translating the SSA2 program into RSSA, we end up with two main
functions:

1) The logical 'main' function from SSA2
2) A function to store globals, eventually calling the 'main' in (1).

The attempt to update this didn't handle the FuncEntry labels correctly:
it gave the new main (2) the same FuncEntry as the SSA main (1).  Since
the SSA main stuck around, and FuncEntrys become labels in the Machine
IL we ended up with duplicate 'FuncEntry's that belonged to different
functions.  Fortunately, the machine type checker caught this error.

The new main function now gets a FuncEntry, which updated in the
RSSA.Program.main recored.