Fix SML/NJ build for Backend.
Add 'Bug' matches to all of the case expressions that scruitinize a
Transfer.t in the RSSA IL.

Commit c536cac introduced a new Transfer.t variant, 'Bug', into the
RSSA IL to accommodate the construction of a fake block to dominate
all of the entry blocks of a function.  However, that commit did not
add 'Bug' matches to all of the case expressions that scruitinize a
Transfer.t.  SML/NJ treats non-exhaustive matches as an error, rather
than a warning, so all such case expressions result in an aborted
SML/NJ build.