Add all entry points to the same (function) class
The original attempt to add support for functions with multiple entry
points tried to set the class of the function for each entry point -- an
error since setFuncClass is a Property.getSetOnce.

The original intention was to map the function to the union of the
classes of the entry points, which is what this change does.  Now any
block reachable (via intra-function transfers) from any of the entry
points of a function will be part of that function's class.

Choosing a union, rather than simply the class of a random entry point,
is important because some entry points may never have control reach a
block that other entry points reach.  This is especially likely to
happen when multiple functions have been coalesced together into one
function earlier in the pipeline.