Change type of 'main' field in {SsaTree,SsaTree2}.Program.T
Change type of 'main' field in {SsaTree,SsaTree2}.Program.T from
"FuncEntry.t" to "{entry: FuncEntry.t, func: Func.t}".  This is the
type of the 'main' filed in Rssa.Program.T and, as discussed in
c536cac, there are a number of instances where accessors wanted both
"main, the collection of blocks" and "main, the entry to start at".