Use single-entry or multi-entry pipeline exclusively.
Now that the multi-entry pipeline produces an unchanged MachineIL
program that is compatible with the codegens, there isn't a strong
reason to resume the single-entry pipeline after the multi-entry
pipeline completes.  Furthermore, the diagnostics that save
information about individual optimization passes sometimes results in
identical file names for both pipelines, overwriting the multi-entry
diagnostics with single-entry diagnostics.  As we begin to port
optimization passes to the multi-entry ILs, it will be important to
access these diagnostics easily.