Remove 'function' field in {SsaTree,SsaTree2,Rssa}.FunctionEntry.T.
While a FuncionEntry.t belongs to a unique Func.t function, it isn't
necessary to record that association in the FunctionEntry.T datatype.
For instance, every Statement.t, Transfer.t, and Block.t also belongs
to a unque Func.t function, but the association is not recorded in the
corresponding datatype.  There were no non-trivial uses of the
'function' field; indeed, in most instances, the 'function' field was
simply copied verbatim from one FunctionEntry.t to a new
FunctionEntry.t.