After duplicating entries, make the callers use the new entries
Another step in exercising the multi-entry pipeline, this pass has been
modified so that callers of the duplicated functions will exercise the
new entry points.

The next step is to have callers pick at random from a list of
replacement entry points, so that both entry points will be exercised
(by different callers).