Updates to conversion of Restore to multi-entry
Commit 48134b gave an initial conversion of Restore to multi-entry.
That commit noted that the major change was how (dummy) entry blocks
are added before calculating the phi functions; these entry blocks
serve as a landing pad for necessary statements.  The minor bug in
commit 48134b is that it only processed dominator trees in the
dominator forest that were rooted by the 'start' block of an entry (in
which case the entry's block as a new root).  Thus, it failed to
process any dominator tree that was not rooted by a 'start' block of
an entry.  Such dominator trees are likely to exist in a multi-entry
functions, where the multiple entries share a common set of blocks.

However, as with commit d139b7, we should take care in the presence of
entries that share a 'start' block.  In such a situation, the entries
that share a 'start' block should each result in an singleton
dominator tree compried of the entry's block.