Add even-odd benchmark.
Add a new even-odd benchmark that exercises mutual tail recursion.
Note that it is important that both "even" and "odd" are called from
the main program; if only "even" is called, then the Contify
optimization will move both "even'" and "odd'" to their one return
point and turn the tail recursion into a local loop.

A forthcoming MergeTailCalls optimization will merge the "even" and
"odd" functions into single function with two entry points; then
IntroduceLoops will eliminate the self-tail calls in favor of a local
loop; then InlineNonRecursive will (most likely) be able to inline the
resulting non-recursive function.  Even if the function is not
inlined, turning the tail calls into a local loop will be a
performance win.