Extend MergeTailCalls pass.
As noted in commit e4de452, it reasonable to build the interprocedural
tail-call graph with both directed edges from callers to callees and
directed edges from callees to callers, in order to find more and
larger strongly connected components to merge.  The requirement for
adding a directed edge from a callee to a caller is that the caller
and callee have compatible "returns: Type.t vector option" and
"raises: Type.t vector option".

Both the original behavior (which only merges mutually tail recursive
functions) and the new behavior described above are provided, the
original behavior as the MergeRecTailCalls pass and the new behavior
as the MergeAllTailCalls pass.  In the SSA optimization pipeline, it
does not seem appropriate to use MergeAllTailCalls before the inlining
passes, as many small functions in tail-call relationships may be
merged into one function that is too large to be inlined.  After the
inlining passes, it does seem appropriate to use MergeAllTailCalls.
However, preliminary benchmarks show that that MergeAllTailCalls can
have a slight negative effect.