Implement SplitEntries pass.
The SplitEntries pass utilizes multi-entry functions to split entry
points that immediately 'Case' on a test argument into an entry point
for each of the constructors, where the constructor arguments are
passed as arguments rather than the constructed value.  Any call to
the original entry point with a manifest constructed value for the
test argument are transformed to call the appropriate split entry
point.

It is not entirely clear where to add SplitEntries to the SSA
optimization pipeline.  It is useful to run SplitEntries shortly
before a RemoveUnused, since SplitEntries may introduce entry points
that are unused.  With this commit, SplitEntries is run twice: once
approximately half-way through the SSA optimization pipeline (after
inlining and the second run of LocalFlatten) and once at the end of
the SSA optimization pipeline (immediately before KnownCase and the
final RemoveUnused).  Preliminary testing shows that the second run of
SplitEntries typical finds many more opportunities to transform calls.