Improve XML sccFuns optimization
When breaking Fun decs into strongly connected components, a singleton
component that corresponds to a non-recursive function may be converted
to a MonoVal or PolyVal dec.

This can improve the efficiency (but not generally not quality) of
subsequent closure conversion by avoiding unnecessary overhead of
apparently recursive functions (e.g., rebinding a self-closure in
TyTransform).