Allow `-duplicate-entries-degree ~1` to mean unbounded duplication
A negative degree is interpreted as duplicating so as to have a
distinct entry for each call site in the program.

To support unbounded duplication, count the number of call sites for
each entry before duplicating.  This count is used as the duplication
degree for the entry for unbounded duplication.  This count is also
used to cap the duplication degree for bounded duplication; this
prevents creating extraneous unused entry points (and lowers the
dependence on an immediately following RemoveUnused).  As a special
case, a entry with count 0 is nontheless duplicated with degree 1,
since the SSA type checker does not accept a function with no entries.