Make main entry implicit in RSSA IR
In SSA and SSA2, we have

  datatype Program.T = {functions: Function.t list,
                        main: {func: Func.t, entry: FuncEntry.t},
                        ...}

but in RSSA, we had

  datatype Program.T = {functions: Function.t list,
                        main: {func: Function.t, entry: FuncEntry.t},
                        ...}

That is, in SSA and SSA2, the list of functions includes the main
function, which is identfied by name in the Program.T type, while in
RSSA, the list of functions does not include the main function, which
is stored separately in the Program.T type.  One reason for this is
that in RSSA, the main function actually corresponds to the globals
from SSA and the variables bound by the RSSA main function are in
scope in the other functions.

In any case, in RSSA, we now have

  datatype Program.T = {functions: Function.t list,
                        main: Function.t,
                        ...}

with the expectation that the main function has exactly one nullary
entry.  This somewhat minimizes the diff with `master`.