Begin updating Inline to work by-entry, rather than by-function
With multi-entry functions, we should inline by entry, rather than by
function.  In particular, MergeAllTailCalls can result in a small
function merged with a large function (while retaining calls to the
small function's entry point elsewhere in the program), but we would
still like to inline the code corresponding to the small function
while not inlining the code corresponding to the large function.
Similarly, with SplitEntries, a specialized entry may reach only a
small portion of the function (and be suitable for inlining).

This commit begins by updating the inlining transformation to work
by-entry, rather than by-function.  The inlining heuristics, though,
remain by-function.