Update leaf{Once,Repeat}{,NoLoop} heuristics to work by-entry
These are simple heuristics, easily extended to work by-entry.  Note
that the call-graph, for leafRepeat{,NoLoop}, is constructed by-entry;
that is, there is an edge from entry E1 to entry E2 if a call of E2 is
reachable from entry E1.  Similarly, whether or not an entry is a leaf
or contains a loop is determined by the blocks reachable from the
entry.  Thus, not all entries in the same function may have the same
properties.