Ininitial update nonRecursive heuristic to work by-entry
This is a non-trivial heuristic and it isn't easy to extend it to work
by-entry.  Recall, that with single-entry, the heurisitic is
effectively:

  fun shouldInline f =
    not (callsSelf f)
    andalso
    (numCalls f - 1) * (size f - small) <= product

with sizes calculated after inlining into f.  Note that the
`numCalls f - 1` effectively allows any non-recursive function with
one call site to be inlined, while the `size f - small` effectively
allows any non-recursive function below a size threshhold to be
inlined.

This commit simply extends the heuristic to work by-entry:

  fun shouldInline e =
    not (callsSelf e)
    andalso
    (numCalls e - 1) * (size e - small) <= product

where `callsSelf e` is true if there is a call to the entry e
reachable from the entry e, `numCalls e` is the number of calls to the
entry e, and `size e` is the size of blocks reachable from the entry
e (after inlining into those blocks reachable from the entry e).

Note, however, that this heuristic can "backfire" in multi-entry.
With DuplicateEntries (especially `-duplicate-entries-degree ~1`,
which introduces an entry for each call-site), the `numCalls e - 1`
can be much smaller than it would be without the DuplicateEntries
pass; it doesn't make sense that more inlining should be enabled just
because there are multi-entries to the same reachable code.  On the
other hand, SplitEntries can introduce specialized entries and it
would be reasonable to allow some of those entries to be inlined even
if the original entry should not be inlined.  For example, consider a
function with an `int list` argument, for which SplitEntries
introduces @nil and @cons entries, in addition to the original @E
entry, where the @nil entry reaches a simple (non-recursive) base-case
computation, while the @cons entry reaches a large and/or recursive
computation.  The @cons entry should not be inlined, but it would be
great to allow the @nil entry to be inlined.  Similarly,
MergeAllTailCalls can result in a small function merged with a large
function (while retaining calls to the small function's entry point
elsewhere in the program), but we would still like to inline the code
corresponding to the small function while not inlining the code
corresponding to the large function.