Update nonRecursive heuristic
Recall that an earlier commit revised the nonRecursive heuristic to
work by-entry as follows:

  fun shouldInline e =
    not (callsSelf e)
    andalso
    (numCalls e - 1) * (size e - small) <= product

Bit, it was noted that this heuristic can "backfire" in multi-entry.
With DuplicateEntries (especially `-duplicate-entries-degree ~1`,
which introduces an entry for each call-site), the `numCalls e - 1`
can be much smaller than it would be without the DuplicateEntries
pass; it doesn't make sense that more inlining should be enabled just
because there are multi-entries to the same reachable code.  On the
other hand, SplitEntries can introduce specialized entries and it
would be reasonable to allow some of those entries to be inlined even
if the original entry should not be inlined.  For example, consider a
function with an `int list` argument, for which SplitEntries
introduces @nil and @cons entries, in addition to the original @E
entry, where the @nil entry reaches a simple (non-recursive) base-case
computation, while the @cons entry reaches a large and/or recursive
computation.  The @cons entry should not be inlined, but it would be
great to allow the @nil entry to be inlined.  Similarly,
MergeAllTailCalls can result in a small function merged with a large
function (while retaining calls to the small function's entry point
elsewhere in the program), but we would still like to inline the code
corresponding to the small function while not inlining the code
corresponding to the large function.

This commit revises the heuristic to work by-entry as follows:

  fun major (f, e) =
    (size e) >= (size f) div 2

  fun shouldInline (f, e) =
    not (callsSelf e)
    andalso
    let
      val numCalls =
        if major (f, e)
          then numCalls e
          else numMajorCalls f
    in
      (numCalls - 1) * (size e - small) <= product
    end

The idea is to identify the "major" entries of a function and treat
their calls as a group.  A major entry is an entry whose reachable
size is at least half the size of the function (considering all
entries); note that more than one entry in a function can be
considered major.

This heuristic avoids the "backfire" with DuplicateEntries, because
all duplicated entries are considered major and all duplicated entries
will use the same `numCalls - 1` value.  Meanwhile, it still enables
inlining the @nil entry and the "small function merged with large
function", because such functions are not considered major.