Install (final) top-level exn handler earlier in Basis Libary eval
Building or running MLton on a new platform sometimes encounters the
following:

    unhandled exception: SysErr
    MLton bug: unhandled exception in Basis Library
    Please send a bug report to MLton@mlton.org.

This comes from the "emergency" top-level exception handler installed.
However, to track down the origin of unhandled exceptions in the Basis
Library, it is useful to compile with `-const "ExnHistory true"`, but
the "emergency" top-level exception handler does not capture or print
exception histories.

This commit moves the installation of the final top-level exception
handler to earlier in the evaluation of the Basis Library.  (Though,
now unhandled exceptions in the Basis Library won't get "flagged" as
such along with a request for a bug report.)