Implement Multi-Return Sig/Struct
Reworked the `Return` structure and signature so that it now expresses a
singular return point, either tail or non-tail. The multiple return
point implementation will be handled by the ssa-tree structures,
treating their returns as a vector of this new Return type.

There are a few TODOs in pending discussion with Fluet. Mostly small
implementation details, these changes are by no means final.